Fixed a bug that caused the palindrom checker to return false for numbers
that were a palindrom because they were declaired as doubles and the
stringbuilder class included a .0

+++ src/checkNumber/CheckNumberTest.java
@@ -3,7 +3,7 @@ package checkNumber;
 public class CheckNumberTest {
 	public static void main(String args[]) {
 		System.out.println("Please input a number you would like the characteristics of to be checked.");
-		double targetNumber = Keyboard.getInteger(); //the number that is checked
+		int targetNumber = Keyboard.getInteger(); //the number that is checked
 		CheckNumber test1 = new CheckNumber(); //test object used to test class
 		System.out.println("The number " + targetNumber + " is " + ((test1.isPrime(targetNumber) == true) ? "" : "not ") + "a prime number.");
 		System.out.println("The number " + targetNumber + " is " + ((test1.isPositive(targetNumber) == true) ? "" : "not ") + "a positive number.");
